Add: modularity in javascript code
