+++ index.html
@@ -15,7 +15,7 @@
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
     />
-    <script defer src="script.js"></script>
+    <script defer src="script.js" type="module"></script>
 </head>
 <body>
     <div class="calculator">
